Struct indexed_vec::IndexedVec

pub struct IndexedVec<T> {
    // some fields omitted
}

IndexedVec is a vector with a unique approach to indices. Once an item is added to the IndexedVec, a unique index is returned. This index is the only way that that element may be accessed, and it is garunteed by the rust typesystem for there only to be one of these indices at one point in time.

This means that we can perform operations that would otherwise be unsafe in a perfectly safe maner. For example, you can grab a mutable reference to an element from an immutable IndexedVec. This is hugely useful in multithreaded environments.

Because the Index is garunteed to exist and point to a valid location in the backing array, the implementation of IndexedVec can also do lookups without bounds checking.

Methods

impl<T> IndexedVec<T>

fn new() -> IndexedVec<T>

Creates a new BoundedArray with a given size.

fn with_capacity(capacity: usize) -> IndexedVec<T>

fn add(&mut self, value: T) -> Index

Adds an element to the BoundedVec.

This function prefers to fill up holes in the array left by removing other items.

fn push(&mut self, value: T) -> Index

Adds an element to the BoundedVec.

This function prefers to add elements to the 'end' of the array before filling holes. It will fill holes if otherwise a resize would be required.

fn get<'a, 'b, 'c>(&'a self, index: &'b Index) -> &'c T

Returns a reference to an element in the array.

fn get_mut<'a, 'b, 'c>(&'a self, index: &'b mut Index) -> &'c mut T

Returns a mutable reference to an element in the array.

fn swap(&self, index: &mut Index, value: T) -> T

Swaps the element at an index, returning the previous value.

fn take(&mut self, index: Index) -> T

Remove the element stored at Index location, returning it.

fn remove(&mut self, index: Index)

Removes the element stored at Index location, dropping it.